refactor(vis_app): Split large files into modular structure
- services/figz/__init__.py: Extract bundle_ops, canvas, constants, layout, panel_ops, preview
- services/gallery_service.py: Extract core, metadata, pltz
- services/plots_service.py: Extract data, files, metadata, render
- services/stats_service.py: Extract context, effects, tests
- services/plot_renderer/element_bboxes.py: Split into element_bboxes/ directory
- urls.py: Split into urls/ directory
- views/api/gallery.py: Extract base, project, utils
- views/api/scitex_editor.py: Extract render, styles
